Restructure locations of yelp reviews test and util classes
